FIx visibility.

+++ Core/src/main/kotlin/org/brightify/reactant/core/ViewBase.kt
@@ -93,5 +93,5 @@ open class ViewBase<STATE, ACTION>(context: Context, initialState: STATE, layout
 
     protected fun <V: View> find(id: Int): V = findViewById(id)
 
-    protected fun Int.toPx() = this * resources.displayMetrics.density.toInt()
+    protected fun Int.toPx() = this * resources.displayMetrics.density
 }
